# Conflicts:
#	msg/de_DE.utf8/cubrid.msg
#	msg/en_US.utf8/cubrid.msg
#	msg/en_US/cubrid.msg
#	msg/es_ES.utf8/cubrid.msg
#	msg/fr_FR.utf8/cubrid.msg
#	msg/it_IT.utf8/cubrid.msg
#	msg/ja_JP.utf8/cubrid.msg
#	msg/km_KH.utf8/cubrid.msg
#	msg/ko_KR.euckr/cubrid.msg
#	msg/ko_KR.utf8/cubrid.msg
#	msg/ro_RO.utf8/cubrid.msg
#	msg/tr_TR.utf8/cubrid.msg
#	msg/vi_VN.utf8/cubrid.msg
#	msg/zh_CN.utf8/cubrid.msg
#	src/base/error_code.h

+++ src/query/string_opfunc.c
@@ -1925,7 +1925,7 @@ db_string_quote (const DB_VALUE * str, DB_VALUE * res)
 	      special_idx.push_back (i);
 	    }
 	}
-      int dest_size = src_size + special_idx.size () + 2;
+      int dest_size = (int) (src_size + special_idx.size () + 2);
       char *result = (char *) db_private_alloc (NULL, dest_size);
       if (result == NULL)
 	{
